vdfvfdvdfvfdvfdvfd

+++ Plant/ViewModels/ProductVm.cs
@@ -24,7 +24,7 @@ namespace Plant.ViewModels
         public string Color { get; set; }
         public int Stock { get; set; }
         public decimal AverageStar { get;set; }
-        public decimal CountProductSell { get; set; }
+        public int CountProductSell { get; set; }
         public int CountFeedback { get; set; }
     }
 }
